Abstract copying a directory to test utils
