+++ test/rename-images.js
@@ -2,29 +2,15 @@ var fs = require('fs')
 var path = require('path')
 var test = require('tape')
 var cleanup = require('./utils/cleanup')
+var copyDir = require('./utils/copy-dir')
 var renameImages = require('../lib/rename-images')
-
-var copyRecursiveSync = function (src, dest) {
-  var stats = fs.statSync(src)
-  var isDirectory = stats && stats.isDirectory()
-
-  if (isDirectory) {
-    fs.mkdirSync(dest)
-    fs.readdirSync(src).forEach(function (childItemName) {
-      copyRecursiveSync(path.join(src, childItemName),
-                        path.join(dest, childItemName))
-    })
-  } else {
-    fs.linkSync(src, dest)
-  }
-}
 
 test('Sorts the images by creation date and renames them', function (t) {
   var imagesDir = path.join(__dirname, 'images')
   var imagesDirCopy = path.join(__dirname, 'images-test')
 
   cleanup(imagesDirCopy)
-  copyRecursiveSync(imagesDir, imagesDirCopy)
+  copyDir(imagesDir, imagesDirCopy)
 
   var images = fs.readdirSync(imagesDirCopy).map(function (image) {
     return path.resolve(imagesDirCopy, image)
